+++ src/i2c.rs
@@ -140,7 +140,7 @@ where
     ) -> Result<I2cMasterDriver<'d, I2C>, EspError> {
         // i2c_config_t documentation says that clock speed must be no higher than 1 MHz
         if config.baudrate > 1.MHz().into() {
-            return Err(EspError::from(ESP_ERR_INVALID_ARG as i32).unwrap());
+            return Err(EspError::from(ESP_ERR_INVALID_ARG).unwrap());
         }
 
         crate::into_ref!(i2c, sda, scl);
@@ -492,7 +492,7 @@ where
         if n > 0 {
             Ok(n as usize)
         } else {
-            Err(EspError::from(ESP_ERR_TIMEOUT as i32).unwrap())
+            Err(EspError::from(ESP_ERR_TIMEOUT).unwrap())
         }
     }
 
@@ -509,7 +509,7 @@ where
         if n > 0 {
             Ok(n as usize)
         } else {
-            Err(EspError::from(ESP_ERR_TIMEOUT as i32).unwrap())
+            Err(EspError::from(ESP_ERR_TIMEOUT).unwrap())
         }
     }
 }
@@ -529,7 +529,7 @@ impl<'buffers> CommandLink<'buffers> {
         let handle = unsafe { i2c_cmd_link_create() };
 
         if handle.is_null() {
-            return Err(EspError::from(ESP_ERR_NO_MEM as i32).unwrap());
+            return Err(EspError::from(ESP_ERR_NO_MEM).unwrap());
         }
 
         Ok(CommandLink(handle, PhantomData))
